requirements.txt
openai
django
python-docx
pydantic
gunicorn
sqlalchemy
pyodbc